ARROW-8261: [Rust-DataFusion] Made limit accept integers and no longer accept expressions.
Also made the type consistent across the project (=usize).

This change is backward incompatible.

The rational for this change is that limit() is almost never called
with an expression in real world cases, since it is used to limit
results before an action.

Closes #7713 from jorgecarleitao/arrow-8261

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>

+++ rust/datafusion/src/table.rs
@@ -44,7 +44,7 @@ pub trait Table {
     ) -> Result<Arc<dyn Table>>;
 
     /// limit the number of rows
-    fn limit(&self, n: u32) -> Result<Arc<dyn Table>>;
+    fn limit(&self, n: usize) -> Result<Arc<dyn Table>>;
 
     /// Return the logical plan
     fn to_logical_plan(&self) -> LogicalPlan;
